# crates/userserver/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "BSCP API",
    "version": "1.0.0",
    "description": "Beorn's Stupid Chat Protocol — user server API. This document is only served by debug builds."
  },
...
  "components": {
    "securitySchemes": {
      "sessionCookie": { "type": "apiKey", "in": "cookie", "name": "bscp_session" },
      "sessionHeader": { "type": "apiKey", "in": "header", "name": "X-Session-Token" }
    },
...
  },
  "paths": {
...
    "/api/admin/config": {
      "get": { "tags": ["admin"], "summary": "Get default storage limit", "security": [{ "sessionCookie": [] }, { "sessionHeader": [] }], "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "object", "properties": { "storage_limit_mb": { "type": "integer" } } } } } } } },
      "patch": { "tags": ["admin"], "summary": "Update default storage limit", "security": [{ "sessionCookie": [] }, { "sessionHeader": [] }], "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "storage_limit_mb": { "type": "integer" } } } } } }, "responses": { "200": { "description": "OK" } } }
    },
    "/api/admin/users/{username}/storage": {
      "get": { "tags": ["admin"], "summary": "Get a user's storage limit", "security": [{ "sessionCookie": [] }, { "sessionHeader": [] }], "parameters": [{ "name": "username", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "OK" } } },
      "patch": { "tags": ["admin"], "summary": "Update a user's storage limit", "security": [{ "sessionCookie": [] }, { "sessionHeader": [] }], "parameters": [{ "name": "username", "in": "path", "required": true, "schema": { "type": "string" } }], "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "storage_limit_mb": { "type": "integer" } } } } } }, "responses": { "200": { "description": "OK" } } }
    },
    "/webhooks/{webhook_id}/{webhook_token}": {
      "post": { "tags": ["webhooks"], "summary": "Incoming webhook endpoint (external callers)", "parameters": [{ "name": "webhook_id", "in": "path", "required": true, "schema": { "type": "string" } }, { "name": "webhook_token", "in": "path", "required": true, "schema": { "type": "string" } }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["content"], "properties": { "content": { "type": "string" }, "username": { "type": "string" }, "avatar_url": { "type": "string" } } } } } }, "responses": { "201": { "description": "Created" }, "404": { "description": "Invalid webhook" } } }
    },
    "/media/proxy": {
      "get": { "summary": "Cached media proxy", "parameters": [{ "name": "url", "in": "query", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "The media bytes" }, "400": { "description": "Missing URL" }, "502": { "description": "Fetch failed" } } }
    },
...
  }
}